refactor(client): FindModule in useModule hook
